refactor: replace UserProfileDropdown with UserAvatar component and streamline settings layout
